+++ cypress/e2e/common/type_invalid_id_in_url.js
@@ -1,0 +1,5 @@
+import { Given } from '@badeball/cypress-cucumber-preprocessor'
+
+Given('I type an invalid dashboard id in the browser url', () => {
+    cy.visit('#/invalid')
+})